+++ HISTORY.md
@@ -1,7 +1,50 @@
 # Snap! (BYOB) History
 
 ## in development:
+* **New Features:**
+    * export script (including dependencies)
+    * export / import sprite-local custom block definitions
+    * added "combinations" primitive to the palette
+* **Notable Changes:**
+    * moved "append", "reshape", "combinations" blocks down one group in the palette
+* **Notable Fixes:**
+    * guard against broken SVG costumes when loading a project
+    * fixed an edge case for slot type inferral
+    * fixed variadic AND/OR reporters library, thanks, Brian!
+* **Documentation Updates:**
+* **Translation Updates:**
+    * German
 
+### 2022-03-15
+* blocks, store, gui: deserialize new format for exported scripts
+* blocks: new "export script" feature (including dependencies)
+* blocks: tweaked "export scripts" for reporters
+* German translation update (for "export script" feature)
+* blocks: added BlockMorph >> toXMLString() api
+
+### 2022-03-14
+* gui, byob: refactored library serialization
+* blocks, byob, gui: new format for exported scripts, under construction
+
+### 2022-03-11
+* blocks: fixed an edge case for slot type inferral
+* objects: added "combinations" primitive to the palette
+* moved "append", "reshape", "combinations" blocks down one group in the palette
+* incremented dev version
+* updated variadic reporters library, thanks, Brian
+* objects: adjusted library-import for sprite-local blocks
+
+### 2022-03-10
+* gui: made dev-warning closeable and added official url 
+
+### 2022-03-09
+* new dev version
+* byob, store, gui: export / import sprite-local custom block definitions, under construction
+* byob, gui: adapted library import dialog to the new format
+* byob: added collecting dependencies for sprite-local custom blocks
+* objects: guard against broken SVG costumes when loading a project
+
+## 7.3.0:
 * **New Features:**
     * variadic associative infix reporters
     * shift-click on STOP button to stop all scenes
@@ -14,9 +57,12 @@
 * **Notable Fixes:**
     * addressed security issues in the JIT-compiler, thanks, xBZZZZ!
     * animation library: fixed pen state for GLIDE, thanks, Brian!
-* **Documentation Updates:**
 * **Translation Updates:**
     * German
+
+### 2022-03-07
+* blocks: refactored experimental export scripts feature
+* prepared v7.3.0
 
 ### 2022-03-04
 * scenes, objects, threads, gui: refactored STOP